+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `laravel-sevdesk-api` will be documented in this file
 
+## 0.1.7 - 2024-02-03
+
+- fix small bug
+- add `allBetween()` to invoice
+- implemented credit note api
+- update documentation
+
 ## 0.1.6 - 2024-01-09
 
 - fix typos in documentation